+++ Changes.md
@@ -53,7 +53,11 @@ Fixes
 - Cycles :
   - Fixed crash rendering a scene containing a CoordinateSystem (#4865).
   - Fixed crash rendering a PointPrimitive with `N` primitive variable (#4876).
+  - Fixed handling of `Varying` primitive variable interpolation (#4849).
+  - Fixed handling of `Constant Color3f` primitive variables.
+  - Fixed handling of indexed primitive variables.
 - PathFilter : Fixed error when selecting a path element from a promoted `PathFilter.paths` plug (introduced in 0.61.13.0).
+- ImageView : Fixed error with display of negative colors.
 
 API
 ---
@@ -137,6 +141,7 @@ Fixes
 
 - NodeEditor : Fixed updated of section summaries when they are changed in the UI Editor.
 - ArnoldColorManager : Fixed broken presets for `color_space_narrow` and `color_space_linear`.
+- ImageView : Fixed error with display of negative colors.
 
 1.0.6.0 (relative to 1.0.5.1)
 =======